+++ src/ui/components/diary/Timeline.stories.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Meta, StoryObj } from '@storybook/react';
 import { Timeline } from '@ui/components/diary/Timeline';
-import { Emotion, Feeling } from '@lib/types';
+import { dummyDiary } from '@lib/const/storybookDummy';
 
 const meta: Meta<typeof Timeline> = {
   title: 'Component/Timeline',
@@ -11,72 +11,6 @@ const meta: Meta<typeof Timeline> = {
 export default meta;
 
 type Story = StoryObj<typeof Timeline>;
-
-const dummyDiary = [
-  {
-    diaryId: 1,
-    feel: Feeling.화남,
-    emotions: [],
-    text: '',
-    createDate: {
-      year: 2024,
-      month: 1,
-      date: 1,
-    },
-  },
-  {
-    diaryId: 2,
-    feel: Feeling.나쁨,
-    emotions: [],
-    text: '다이어리 내용',
-    createDate: {
-      year: 2024,
-      month: 1,
-      date: 1,
-    },
-  },
-  {
-    diaryId: 3,
-    feel: Feeling.보통,
-    emotions: [Emotion.뿌듯한, Emotion.행복한, Emotion.기대되는, Emotion.편안한, Emotion.피곤한],
-    text: '',
-    createDate: {
-      year: 2024,
-      month: 1,
-      date: 1,
-    },
-  },
-  {
-    diaryId: 4,
-    feel: Feeling.좋음,
-    emotions: [Emotion.뿌듯한, Emotion.행복한],
-    text: '다이어리 내용',
-    createDate: {
-      year: 2024,
-      month: 1,
-      date: 1,
-    },
-  },
-  {
-    diaryId: 5,
-    feel: Feeling.행복,
-    emotions: [
-      Emotion.뿌듯한,
-      Emotion.행복한,
-      Emotion.기대되는,
-      Emotion.편안한,
-      Emotion.피곤한,
-      Emotion.슬픈,
-      Emotion.부담되는,
-    ],
-    text: '다이어리 내용',
-    createDate: {
-      year: 2024,
-      month: 1,
-      date: 1,
-    },
-  },
-];
 
 export const Diary1: Story = {
   render: () => (
